Multiple Changes
Changed which arrays are sorted or not because of how the
MULTIBUILD_VARIANTS array is handled by the multibuild.eclass when we
want to get the best, which is assumed to be the last element in the
array.

Moved and renamed the postgres-multi-get_impls() from
postgres-multi.eclass to postgres.eclass as postgres-get_impls() to
not reapeat myself as I actually need the same behavior for
postgres_single_slot_pkg_setup().

Use MULTIBUILD_VARIANTS instead to get the best implementation.

And other minor stuff.