Fix Compilation
Get the 9.0 patches instead of the 9.1 patches. Since the makefiles
automatically detect whether or not to use/enable XML support, we don't
need to explicitly state it ourselves. And, because the build order is
the same whether or not we are building the server, skip that test.

And, get rid of those damn extra pluses sed decided to put in.