Improve Live Ebuild and Simplify Compilation
There were several holdovers from the old ebuilds that were entirely
unnecessary for quite some time. They've been removed.

The base.patch modifies the Makefiles so that we don't have to worry
about whether we're building the server or not. The server-related
targets won't be made when the 'server' USE flag is disabled.

Dropped 'dohtml ...' in favor of 'docinto html; dodoc ...' per
gentoo-dev list conversation. [1]

Include a sample psqlrc file instead of just keeping the
/etc/postgresql-${SLOT} directory to make it a bit easier to figure out
what to do with it when you don't have the server installed.

Corrected the src_install() on postgresql-9.2.5.ebuild as it wasn't
properly unified.

Include the previously ignored 'man 3 ...' pages.

[1] http://www.gossamer-threads.com/lists/gentoo/dev/291541