Wow...just...wow....
This was quite possibly the most infuriating set of fixes I've ever
done. I am simply amazed that there were NO bugs filed against
www-apps/rt except to bump the version.

The reconfig script was broken. The ebuild was broken. Even the
recommended Apache config was broken. The guide we linked to was
severely out of date. Then there's the install stripping the prefix
along with the trailing slash so all the directory paths are relative
instead of absolute.

This has filled my last 24 hours with many head desk, expletive
excreting, tearful moments as I tried to grasp at small glimmers of
sanity only to find that I was hallucinating.

On the plus side, I've discovered that I'm rather resolute. So...there's
that.

Package-Manager: portage-2.2.8-r2
Manifest-Sign-Key: D1BBFDA0