Remove server and doc use flags
Build against upstream's HEAD can result in breakages, which would
require more effort by the ebuild maintainers than the compiler. So,
we'll just require an extra minute and a few more megs of space of the
user.

We'll also unconditionally build the docs because they'll most likely
not be available online.