Make make respect envvars
Despite being told where the includes and libs are for each slot, the
src/sql/Makefile uses pg_config to find the pgxs file. But, the
Makefile didn’t actually look for any outside input as to where it
should actually try to find pg_config or look to see if PG_CONFIG was
already set.

A couple question marks resolves this and allows us to install the
required extensions for each slot.