Avoid File Collisions
A bit of a hack to work around file collisions that really aren’t a
collision. The slot-specific symlinks (e.g, psql96) were created by
the postgresql.eselect module, but we’ve move that into the
ebuild. This has the side effect of causing file collisions.

To solve this, src_install creates symlinks with an alternate name,
and pkg_preinst cleans out the old eselect generated symlinks and
renames the ebuild generated symlinks to take their place.