dev-db/pgpool2: Bump to 3.7.4, configure.ac fixes
Bumped to 3.7.4. Added some additional fixes for the configure script
to properly handle when memcached is enabled. libmemcached may have
been built against the sasl library which may be threaded. As pgpool’s
build system didn’t do any discovery on how to handle threading, it
would report libmemcached as not being installed when it was.

Fixed the memcached option so that no longer need to be given
a path instead using pkg-config to determine how to link against the
library and where to find the header.

Fixed the pam option so that giving it --without-pam doesn’t actually
enable it.

Package-Manager: Portage-2.3.40, Repoman-2.3.9